Ok, it's time to bugfix since the release get much more important ;-)
Bugfix:
Added missing layer to the continerViews menu items.

Enhance:
Added layer, extra, order, icon, menu and for attributes to the addMenuItem directive.
This makes the addMenuItem useable for not only the AddMenu class in the IDefaultLayer

Moved:
The addMenuItem directive get moved to the menumeta.py where was added from srichters
menu refactoring

Test:
Changed and added more tests

git-svn-id: svn://svn.zope.org/repos/main@29569 62d5b8a3-27da-0310-9561-8e5933582275