Use a special 3.4a1 in Zope3.
Zope3 doesn't include all of ZODB, and in particular skips
the old Peristence module.  But one of the ZODB tests
(checkOldStyleRoot) has an obscure dependence on Persistence
(buried in a hand-constructed pickle), so that test fails in
Zope3's idea of what ZODB is.  The new tag is 3.4a1 with that
test changed to tolerate the absence of Persistence.


git-svn-id: svn://svn.zope.org/repos/main@29852 62d5b8a3-27da-0310-9561-8e5933582275