Temporarily point the ZODB code at current 3.4 branch.
That contains some of the changes from Jim's ZClasses branch.
The other changes still need to be merged into Zope, but require
using current ZODB code.  Pointing at the 3.4 branch allows for
easy changing of both ZODB and Zope until all of this works.

A proper tag for ZODB will be made later.


git-svn-id: svn://svn.zope.org/repos/main@29892 62d5b8a3-27da-0310-9561-8e5933582275