Fix Lennart's failing test.
Lennart believed that he had to monkey patch Zope3's StandardMacros
because it was a new-style class and thus couldn't be registered as
a view from Five (which is quite right). This fact gave him problems
with forms, because certain widgets require the use of the @@form_macros
view, which is based on StandardMacros.

My reasoning, as I've stated many times (e.g. in
http://codespeak.net/pipermail/z3-checkins/2005q1/000677.html) was that
we simply have to provide our own non-new-style (IOW, ExtensionClass-based)
@@form_macros view.

This is what this patch does. Nothing more. The test passes
(Yes, Lennart, I'm having my parade. Please let me :))


git-svn-id: svn://svn.zope.org/repos/main@52619 62d5b8a3-27da-0310-9561-8e5933582275