Expose a bug in PageTemplateResources:
Publishing a resource would fail with a IndexError "tuple index out of range"
in Zope-2's BaseRequest code that calls the __browser_default__ on the
resource instance.

The __browser_default__ method returns an empty tuple wher I guess it should
return the name of the method to call on the resource instance to make it
render the template.



git-svn-id: svn://svn.zope.org/repos/main@53210 62d5b8a3-27da-0310-9561-8e5933582275