Fixed a bug in savepoint rollback.  It's not enough to rollback just the savepoint being rolled back because later savepoints might involved data managers that hadn't joined when the savepoint being rolled back was created.
Now, when a data manager joins and we have savepoints, we create a
data manager savepoint for the new data manager and add the
datamanager savepoint to all previous transaction savepoints.  Note
that this data manager savepoint can be a special savepoint that just
calls abort on the data manager when it is rolled back.


git-svn-id: svn://svn.zope.org/repos/main@30165 62d5b8a3-27da-0310-9561-8e5933582275