Hundreds of temp files (under .svn/ directories) are created in read-only mode, presumably by SVN, when running the tests. Windows won't allow deleting a read-only file (well, at least not via os.remove/os.unlink).
So, a new utils.py introduces a rmtree_force() function, which
walks the tree marking files writable first, and then calls
shutil.rmtree().  app.py is changed to use it.

This fixes 4 of the test failures on Windows.  There are still
lots of other direct calls to shutil.rmtree in this code, though.


git-svn-id: svn://svn.zope.org/repos/main@30210 62d5b8a3-27da-0310-9561-8e5933582275