Fixed a bug/missfeature. Split transition-ordering specifications were ignored.  We were basing evaluation of transition conditions on transition-definition order.  We should, instead, base order of evaluation of transition conditions on the order given in splits.  The spec is a bit vague in this area as it suggests that transition restrictions are optional and that there can be multiple, which doesn't make much sense.
Also changed the strategy for setting up transition information. We
now compute activity incoming and outgoing transitions as we build up
the model, rather than waiting until we compute the start node.

Added an attribute to teh piblic process-definition interface.  This
was already used in the documentation.

Simplified the element-handling logic in the xpdl code.


git-svn-id: svn://svn.zope.org/repos/main@30239 62d5b8a3-27da-0310-9561-8e5933582275