checkUndoInVersion():  Add an optional `hook` argument.
This gets called "in the middle" of the test, if specified.
ZRS 1.5 uses this to pass a callback that arranges to start a
ZRS secondary then.  ZRS had its own copy of this test, but
it's a miserable & messy test, and the copy it had failed in
5 different places when using ZODB 3.4 (it had gotten way of
synch with changes since ZODB 3.2).

Also removed all traces of the bizarre _x_dostore() method.
Not sure what that was about, but if the comments were right
we don't care about ZEO 1.0 anymore.


git-svn-id: svn://svn.zope.org/repos/main@30344 62d5b8a3-27da-0310-9561-8e5933582275