make auto-generated forms unicode aware. In order to do that, a _decode method was copied from zope3's zope.publisher.browser.BrowserRequest. In zope3's zope.publisher it is called for every form field that is not an HTTP header or a CGI name. We mimic the same in the EditView/AddView classes. The processInputs() method needs to be called explicitly from the edit/add form PageTemplate though.
In addition to that, the encoding of the browser page needs to be set.
In Zope3, a view delivers unicode and zope.publisher will send to the
browser the encoding which fits best to what was requested originally.
Zope 2 sadly doesn't know this. Instead it'll see if the response's
Content-Type header was set with a charset parameter and use that, otherwise
it defaults to iso-8859-1 which is not suitable for unicode-awareness
at all. Therefore, EditView/AddView take over the work that normally
zope.publisher does and negotiate the best encoding suitable for the
form and the browser and set the Content-Type header accordingly.
That way ZPublisher will encode correctly.


git-svn-id: svn://svn.zope.org/repos/main@54662 62d5b8a3-27da-0310-9561-8e5933582275