Merge fixes to namechooser from the trunk:
New Revision: 12840
Fix a bug in the objectmanager NameChooser and added a doctest
for it. This code must have been totally untested, in fact
nothing else in Five.browser.adding has a test *shrug*.
svn blames faassen *wink*

New Revision: 12860
jazz up the name chooser again. This time the handling of unicode names
was improved. NameChooser will now convert unicode names to strings
(ObjectManager wants names to be strings) and checks if they're only
ASCII.  Also, the NameChooser.chooseName method is is *always* invoked
now so that we definitely end up with a valid strin-based name.
This is a prerequisite for the unicode-ification of forms.

Also, improved error messages that now give reasons why the name was bad.


git-svn-id: svn://svn.zope.org/repos/main@54675 62d5b8a3-27da-0310-9561-8e5933582275