Mrege rev 30658 from 3.4 branch.
checkCrossDBInvalidations:  make this willing to wait longer in all,
and increase the sleep time per failing loop iteration.  The hope is
this will allow the test to pass on slower and/or "too busy" machines.
I can't test that, though, since I couldn't get the test to fail at
all on a Windows or a Linux box.


git-svn-id: svn://svn.zope.org/repos/main@30659 62d5b8a3-27da-0310-9561-8e5933582275