Removed the `key2size` OIBTree.
Maintaining this is a significant expense; the code didn't
make real use of it; and every time I've tried to exploit it
I've ended up with grossly too-large simulated hit rates.
Would probably take another day to figure out exactly why that
is, and the simulated rates are "good enough" now without it.


git-svn-id: svn://svn.zope.org/repos/main@37563 62d5b8a3-27da-0310-9561-8e5933582275