Primarily a checkin about removing XXXs from zope.app.form.
Fixed disabled test for file widget; in the process made some of the functional
tests more careful.

Ripped out unavailable (and apparently rejected) 'usage' argument in
schemadisplay factory.

Ripped out unused 'menu' argument in many form factories.

Changed appropriate XXX marks to TODO.

In a couple of removed XXX comments, someone, possibly srichter, asked:

    What about the __implements__ of the bases?

The __implements__ of the bases should all be part of the created type.  Not
sure from the question if that was what the writer expected or not.



git-svn-id: svn://svn.zope.org/repos/main@37569 62d5b8a3-27da-0310-9561-8e5933582275