Fix #384: Handling utility permissions
When defining utilities in ZCML with permissions, checker objects were
saved on the utilty's __Security_checker__ attribute. This caused
problems if the component was registered more than once, or if the
component was a class.  Now a proxy is used to hold the security
checker, so that the checker is specific to a particular registration.


git-svn-id: svn://svn.zope.org/repos/main@37586 62d5b8a3-27da-0310-9561-8e5933582275