Merge rev 37713 from 3.4 branch.
Plug leaks in the ZEO client cache.

ClientCache._evicted():  When deleting the last range of
non-current tids for an oid, remove the list from the
noncurrent dict instead of leaving an empty list sitting
there forever.  This also required adjusting the side-
effect dance in ClientCache._remove_noncurrent_revisions().

FileCache._makeroom():  This was the major leak -- it
neglected to remove an evicted object's Entry from the
key2entry dict.


git-svn-id: svn://svn.zope.org/repos/main@37714 62d5b8a3-27da-0310-9561-8e5933582275