Implements the ordering of before commit subscribers :
Compatibility notes : 

 o getBeforeCommitHooks() remove the hook's order values so it returns
the same data structure as before

 o beforeCommitHook() and the order argument.

   The issue here is that if the order argument is extracted from the
**kws then we would have to remove this argument before storing the
hook information and thus the order argument name will not be usable
from the __call__() method of the hook itself.

   A beforeCommitHookOrdered(hook, order, *args, **kws) has been
created taking care of the insertion of the hook given the order. The
beforeCommitHook() method just call beforeCommitHookOrdered() with 0
as default value. Here, the goal is to provide backward compatibility
and a not an ugly implementation extracting the order from kws and then
deleting the entry when extracted before storing.

Notes about sorting and inserting : 

 - bisect.insort() can't be used since insort() will sort on the whole
tuple element which is no what we wanna do.

 - an IOBTree use for the data structure will be a little bit overkill
and will complicate the implementation of the hooks that can call
other hooks at execution time. Plus de persistency is not necessarly in here.



git-svn-id: svn://svn.zope.org/repos/main@37789 62d5b8a3-27da-0310-9561-8e5933582275