Merge rev 38114 from trunk.
Collector 1873.

update_from_seq():  If PySequence_Check(seq) returns true,
go on to check for the existence of "iteritems" when deciding
whether the input is or is not "a sequence".  Alas,
PySequence_Check() does return true for PersistentMapping and
PersistentDict instances, so that they couldn't be used as
arguments to BTree/Bucket construction or update().

This is nasty type-sniffing, but it was before too.  There's
no way to make such stuff bulletproof, so I'm settling for
incremental improvement.


git-svn-id: svn://svn.zope.org/repos/main@38115 62d5b8a3-27da-0310-9561-8e5933582275