BTree_setdefault():  Rewrote to look more like Python's dict.setdefault implementation, and to use 4-space indents like the rest of the file.  This fixes some endcase bugs too.
Alas, treating the default "key missing value" of None
correctly reveals that this isn't a sane policy for
II, OI, or IF BTrees:  they can't have None as a value.

testSetdefault():  Rewrote to reflect correct treatment of
a missing default, and added an XXX noting that the repaired
test fails now for II, OI, and IF trees.

Note:  Buckets need to grow a setdefault() implementation
too.


git-svn-id: svn://svn.zope.org/repos/main@38146 62d5b8a3-27da-0310-9561-8e5933582275