Merge the alienoid-btrees_pop branch.
This adds .pop() methods to BTrees and Buckets.

Alas, zpkg changed since the branch was made, and it
wasn't (obviously) possible to build the branch anymore,
so much of this code didn't actually appear on the
branch.  The C code was changed to simplify and purge
non-ANSI gcc extensions, and the new test cases beefed
up a little.


git-svn-id: svn://svn.zope.org/repos/main@38270 62d5b8a3-27da-0310-9561-8e5933582275