I am making this branch to save the state of pageletchooser, skin tools and the boston skin.
I have reinterpreted the zope.app.pagelet design to be much more simple, 
and I want to check it in, so it can be used by others. However, this 
reinterpretation shakes up the boston skin quiet a bit, and I am not even 
sure if the pageletchooser is still needed. Thus, I will remove them from 
the trunk until we reimplement their functionality using the new design.



git-svn-id: svn://svn.zope.org/repos/main@38492 62d5b8a3-27da-0310-9561-8e5933582275