Putative fix for Collector 1912.
This a port from the Zope(2) trunk, where it applied
to OFS.Image.File.  The Zope3 code in zope/app/file/file.py
appears to be a rewrite of that.

File._set_data():  This uses savepoints just to get oids assigned
and to free memory, with no possibility of rollback.  Therefore
"optimistic" savepoints should be used:  there's no reason here to
insist that any data managers involved support rollback.

If this should be ported to Zope3 branches, I'm not clear on
what they are, so someone else should do that.


git-svn-id: svn://svn.zope.org/repos/main@38759 62d5b8a3-27da-0310-9561-8e5933582275