The standard Zope 2 containers now send events. <five:sendEvents> is removed, and replaced by more <five:containerEvents> and <five:deprecatedManageAddDelete>
This is a work in progress, some XXX remain, notably:
 - add event listeners for _notifyOfCopyTo and _postCopy
 - add a subscriber for manage_afterClone emulation
 - add subscribers for executable ownership and Owner role fixups
More doc will also be included.

Finally, this includes a "transitional" phase, which will only be used
during development but will be removed when things have been suitably
tested with Zope and CMF, and deprecated classes have been converted or
identified.



git-svn-id: svn://svn.zope.org/repos/main@60727 62d5b8a3-27da-0310-9561-8e5933582275