Boston skin --> work in progress
TODO: reimplement menu/menu-item
Right now the menu implementation can't be used for 
register submenus on menu items where get registred in 
the view or page directive. Because the view or page directive
don't register a menu where we can use for register submenus.

The submenu implementation can only be used if we register
menu and submenu out of the box in one complex directive.

git-svn-id: svn://svn.zope.org/repos/main@39683 62d5b8a3-27da-0310-9561-8e5933582275