Fix acquiry from less local sites.
Basically, both the FiveSiteManager and the SimpleLocalUtiltiyService were not
correctly delegating stuff to the next site manager or utilty service, respectively.
This has been fixed now. A ``next`` attribute was introduced to the
IRegisterUtilitySimply interface that represents the next whatever. It is
heavily used by {query|get}NextService(s) which are Zope 3 API functions
for acquiring the next whatever. (The next whatever does not necessarily
have to be the global whatever, just a less local whatever).

Also fixed a Zope 2 Acquisition problem in the utility service that didn't
allow a more local utility to shadow a less local utility registration.

We need tests for the FiveSiteManager. test_sitemanager.py does not test it
all and should be renamed.


git-svn-id: svn://svn.zope.org/repos/main@60931 62d5b8a3-27da-0310-9561-8e5933582275