This is an intermediate checkin, to get the testbrowser code closer to our desired end-state.
Positive changes: we are using the current version of the testbrowser dependencies.  We are not hacking them in with sys.modules (they no longer could be, actually).  We are not hacking the dependencies any more.  The goBack bug should be fixed.  The dependencies should pass all their tests (in their own packages, not yet imported).  The memory problem Stephan found and fixed in our hacked up copies should still be fixed, but now using the dependency code that we are not maintaining.

Negative changes: we have some modules (not packages) in the src directory; Fred and Jim agree that's the best we can do ATM, but we're hopeful that some work that Philipp is doing can make this prettier.  The tests for the dependencies are still not included.  The dependencies now appear to themselves have a dependency on Python being built with ssl support.  These issues are going to be discussed further and hopefully addressed.


git-svn-id: svn://svn.zope.org/repos/main@39818 62d5b8a3-27da-0310-9561-8e5933582275