Rethink strategy for dealing with top-level modules:
We don't put them in Dependencies/<depname>, but we put them in a
Modules/ directory that is top-level to the distribution and a sibling
to Dependencies. Distutils likes to have all top-level modules
(py_modules parameter) to be in one directory, which is incidentally
the one specified by package_dir['']. zpkgsetup now satisfies distutils
in both of those ways.

Now, when dealing with a ModuleComponent, zpkg doesn't create a setup.py,
setup.cfg under Dependencies/<depname> anymore. Instead it creates a
MODULE.cfg, a very simple configuration file telling zpkgsetup the name
of the module. It will then try to find it under Modules/ where zpkgtools
will have put it.

This approach works well so far; it's only that I'm getting a weird error now
("error: Is a directory") when running python install.py build; in particular
it occurs in the build_scripts subcommand.  Unfortunately the error doesn't
occur in python (otherwise I'd have a nice traceback). I'm further
investigating.


git-svn-id: svn://svn.zope.org/repos/main@40096 62d5b8a3-27da-0310-9561-8e5933582275