After Commit Hooks branch
Sometimes, applications want to execute some code after a transaction
is committed. For example, one might want to launch non transactional
code after a successful commit. Or someone might want to launch
asynchronous.  A post-commit hook is available for such use cases: use
addAfterCommitHook(), passing it a callable and arguments.  The
callable will be called with a Boolean value representing the status
of the commit operation as first argument (true if successfull or
false iff aborted) preceding its arguments at the start of the commit
(but not for substransaction commits).

See http://mail.zope.org/pipermail/zodb-dev/2005-September/009233.html
for the discussion that occured related to this.



git-svn-id: svn://svn.zope.org/repos/main@40908 62d5b8a3-27da-0310-9561-8e5933582275