Merge rev 41063 from 3.6 branch.
_ConnectionPool._reduce_size():  when forgetting a Connection
due to exceeding pool_size available connections, clear its
cache right away.  Because such a connection can never be in
the open state again, hanging on to resources in its cache is
just wasteful.  This was reported as "a problem" on zodb-dev
recently, although it's unclear how the poster got into a state
where it mattered so much.


git-svn-id: svn://svn.zope.org/repos/main@41064 62d5b8a3-27da-0310-9561-8e5933582275