Modified the DAV widgets to handle the rendering and parsing of all webdav properties to and from Python xml.dom elements.
Clarified some of the fields in the DAV schema in order to
take advantage of the new widgets. Also reworked a big chunk
of PROPFIND to take advantage tof the new widgets.

Also added a basic implementation of the COPY, LOCK and
UNLOCK methods.

Make the ILockStorage utility in zope.app.locking addable
through the site management interface.


git-svn-id: svn://svn.zope.org/repos/main@41259 62d5b8a3-27da-0310-9561-8e5933582275