Pain, pain, and more pain.
The build-the-installer process now completes, and running
the installer creates something that may or may not be a
working Zope.  It starts fine as a Windows Service, and at
least "looks like a Zope" ;-)

Nothing in the build-the-Windows-installer process here
uses anything in the root of the inst/ directory anymore.
Instead the WinBuilders zope.mk builds Zope all by itself,
using the Python created by python.mk.  Everything these
used to use in the root of the inst/ directory was so out
of whack with current reality that there was no point even
trying to reverse-engineer what it thought it was doing.

Note that comments in the code highlight what look like
bugs in distutils, and in the Windows xcopy command (that
last one took hours to track down -- sheesh).


git-svn-id: svn://svn.zope.org/repos/main@41291 62d5b8a3-27da-0310-9561-8e5933582275