for alecm:
Changed Traversable to call the original __bobo_traverse__ method before 
attempting to use attribute access to find the desired object.  This ensures 
that the monkeypatched object retains the same behavior as the original 
object during traversal.  Also, added checks to see if the request is a 
WebDAV request in which case attribute lookup is performed without 
acquisition (see BaseRequest.traverse for why this is important).



git-svn-id: svn://svn.zope.org/repos/main@65372 62d5b8a3-27da-0310-9561-8e5933582275