Add in checkpoint changes: huge refactoring, with significant change to internal data structure and no migration, so existing instances will need to be trashed (or you'll need to try and convince me why I should write a migration for you).
Now the central piece of the package is a relationship index, which is an N-way relationship index with a lot of very powerful features.  The relationship container now is a comparitively constrained use of the relationship index.  The relationship index can be used standalone or within a catalog.

Also added a listcontainer variety of the relationship container: still needs tests.


git-svn-id: svn://svn.zope.org/repos/main@41745 62d5b8a3-27da-0310-9561-8e5933582275