Forward port from 3.3 branch:
------------------------------------------------------------------------
r68201 | philikon | 2006-05-21 13:23:41 +0200 (Sun, 21 May 2006) | 5 lines

The default Context implementation wasn't fully complying with ITALExpressionEngine,
it was missing the 'translate()' method.  THus, when using it without any customizations
(zope.app and Zope2 do customize the Context, so it's not a problem there), you'd
get an error.  A test in zope.pagetemplate exercises this now.



git-svn-id: svn://svn.zope.org/repos/main@68203 62d5b8a3-27da-0310-9561-8e5933582275