Pluggable Traversers
This pacakge was originally developed for SchoolTool to satisfy the gods 
of SchoolTool (i.e. POV), which hate the '++namespace++' notation. :-)

From the README:

Traversers are Zope's mechanism to convert URI paths to an object of the
application. They provide an extremly flexible mechanism to make decisions
based on the policies of the application. Unfortunately the default traverser
implementation is not flexible enough to deal with arbitrary extensions 
(via adapters) of objects that also wish to participate in the traversal 
decision process.

The pluggable traverser allows developers, especially third-party developers,
to add new traversers to an object without altering the original 
traversal implementation.



git-svn-id: svn://svn.zope.org/repos/main@69480 62d5b8a3-27da-0310-9561-8e5933582275