Fix Zope3-dev issue #673: Bug in configure script regarding python version check
The problem was that when target version of Python was 2.4.3 and acceptable versions
were 2.4.1 and 2.4.2, a newer Python version such as 2.4.4 would not be accepted
by the configure script.

We "fix" this for now by adding possible future Python versions in the acceptable
list.  This isn't a real fix of configure's Python detection algorithm, but as an
XXX comment in confiugre.in says, perhaps the whole algorithm isn't necessary at
all... Either way, better packaging tools are on the Zope horizon (setuptools, ...).

Tested this by getting rid of all Python executables on my $PATH and adding a
"fake" Python interpreter somewhere in the search path:

  #!/bin/sh
  echo "2.4.4"

This would break before, now only issues a warning.


git-svn-id: svn://svn.zope.org/repos/main@69593 62d5b8a3-27da-0310-9561-8e5933582275