Added language package
The package contains a local negotiator,
some base classes with a i18n switch adapter
and a library for session handling support.

This packages allows you to implement simply
i18n aware content types and configure the 
server (negotiator) for handling language 
policies. Such a policy could be, that the server 
will lookup for a fix language or language settings
set on a session or the setting from the browser.

See for more information inthe different README.txt 
files or in the included tests

git-svn-id: svn://svn.zope.org/repos/main@69715 62d5b8a3-27da-0310-9561-8e5933582275