Introduce ModuleGrokker and priority for grokkers. This allows us to get rid of the last bits in _grok that have to do with actual grokking and move it over to meta.py.
ModuleGrokkers get executed once per module. This can then be used to
register filesystem templates, or various decorators, or static files if
the module happens to be a package.

Grokkers get a priority now. The higher the priority, the earlier in the
grokking process they're executed. We use this to make sure that some
grokkers get executed before others.


git-svn-id: svn://svn.zope.org/repos/main@71541 62d5b8a3-27da-0310-9561-8e5933582275