Added a (hysterically poorly named) writelines method to allow iteratibles to be passed to connections for output.
Added a new ConnectionHandler method to allow an implementation to
report errors to an application asynchronously. The main use case for
this is in handling errors from iterators passed to writelines.

Added a blocking I/O API to simplify writing some kinds of
applications.

Finished transition to rule that setHandler can only be called from
callbacks by eliminating some unnecessary locking. (The rules for
allowable callbacks need to be fleshed out a little bit more to deal
with the special case of connections created without a callback, for
example using blocking.connect.


git-svn-id: svn://svn.zope.org/repos/main@71766 62d5b8a3-27da-0310-9561-8e5933582275