Make test.py -s subdir pay attention to the directory -> package mappings.
The goal is to make test.py -s src/zope/subpackage be an alternative
way of specifying -s zope.subpackage, if I'm in the root of a Zope 3 checkout.
Why is that useful?  This way you can use shell tab-completion to specify
a package name.  The current "backwards compatibility" code that replaces
slashes with dots is not sufficient for that.

Merged from the filter-by-dirs branch with

  svn merge -r 71985:71986 svn+ssh://svn.zope.org/repos/main/zope.testing/branches/filter-by-dirs .



git-svn-id: svn://svn.zope.org/repos/main@72031 62d5b8a3-27da-0310-9561-8e5933582275