Implement ZCMLLayer.tearDown.
As a result, the test runner no longer needs to run functional tests of all but
the first layer in a subprocess, and you get progress indication and also can
use pdb there.

All tests (test.py --all) pass with this change.  I've also checked the
contents of the adapter registry, the logging module configuration, the
contents of the transaction, and the list of loggers registered with the Python
logging framework.  All these are not changed by setting up and tearing down
test layers.



git-svn-id: svn://svn.zope.org/repos/main@72143 62d5b8a3-27da-0310-9561-8e5933582275