Add a --checkers option.  You pass a Python module name (e.g. zope.testing.eventcheckers).  You can pass more than one --checkers argument.
The Python module must define a function named test_checkers() that returns
a list of objects.  These objects can define the following methods:

   startTest(test)
   stopTest(test)
   startLayer(layer)
   stopLayer(layer)

The checkers can inspect the application state before and after a test (or
layer) and complain if the test/layer changed something but did not clean up
after itself.

What's missing in this branch:

   - tests for the feature
   - documentation
   - sample checkers:
        - look at logging handlers
        - look at zope.event.subscribers
        - look at data managers registered with the transaction
        - look at the list of component registrations



git-svn-id: svn://svn.zope.org/repos/main@72145 62d5b8a3-27da-0310-9561-8e5933582275