Merged changes from branch.
The API has changed !

The metadata for the table definition allows an explicit engine name as
paramter. Change the use of the metadata like this :
 z3c.zalchemy.metadata -> z3c.zalchemy.metadata()
If you want to assign the table to a named engine provide the name as
a paramter to metadata().
This change allows to have two or more tables with the same name and
different structures to be used in different databases.

Demos have been updated by Michael Bernstein.



git-svn-id: svn://svn.zope.org/repos/main@72193 62d5b8a3-27da-0310-9561-8e5933582275